[editor] to fix the grave uuid we need to, sleep is required
[editor] Fixed Gravzmo, Rapido and the rythm endiabled; Renamed ImGuiToastType to ENotifType
ratz papapapalala

+++ CherryCrisis/CherryEngine/include/uuid.hpp
@@ -7,13 +7,13 @@
 class CCENGINE_API CCUUID 
 {
 private:
-	uint64_t m_uuid;
+	uint32_t m_uuid;
 public:
 	CCUUID();
-	CCUUID(uint64_t uuid): m_uuid(uuid) {}
+	CCUUID(uint32_t uuid): m_uuid(uuid) {}
 	CCUUID(const CCUUID& uuid) { m_uuid = uuid.m_uuid; }
 
-	operator uint64_t() const { return m_uuid; }
+	operator uint32_t() const { return m_uuid; }
 };
 
 namespace std 
@@ -23,7 +23,7 @@ namespace std
 	{
 		std::size_t operator()(const CCUUID& ref) const 
 		{
-			return hash<uint64_t>()((uint64_t)ref);
+			return hash<uint32_t>()((uint32_t)ref);
 		}
 	};
 }
